requirements.txt
asgiref==3.8.1
Django==5.2
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
dnspython==2.7.0
email_validator==2.2.0
idna==3.10
python-dotenv==1.1.0
sqlparse==0.5.3
tzdata==2025.2